fix : parameter로 넘겨주던 ref => 없애기
